get cards start update card add koin

+++ app/src/main/java/com/example/cardpaymentapp/entity/CardModel.kt
@@ -7,6 +7,6 @@ data class CardModel(
     val id: String = UUID.randomUUID().toString(),
     val name: String,
     val type: CardType,
-    val balance: Long,
+    val balance: Int,
     val validUntil: String
 )
